Variable Star
A program in C intended to be used to open and read a file containing known star classes, comparing those values against input data. Chi-squared values are calculated, with results stored in an array. Program will iterate over the array to find the three smallest values. Red, green, and blue LEDs (on a PCduino) will flash to indicate the quality of the matches with a series of actions.